+++ src/main/java/com/billit/user_service/account/domain/repository/BorrowAccountRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface BorrowAccountRepository extends JpaRepository<BorrowAccount, Long> {
     List<BorrowAccount> findAllByUserBorrowIdAndIsDeletedFalse(Long userBorrowId);
-    Optional<BorrowAccount> findByIdAndIsDeletedFalse(Long id);
+    Optional<BorrowAccount> findByIdAndIsDeletedFalse(Integer id);
     Optional<BorrowAccount> findByAccountNumberAndIsDeletedFalse(String accountNumber);  // 추가
     boolean existsByAccountNumberAndIsDeletedFalse(String accountNumber);
 }
